conda_requirements.txt
aiohttp
black
flake8
ipython
mypy
mypy-protobuf
numpy
pip
protobuf
python
pytorch
cpuonly
pyro-ppl
tensorboard
uvloop
